fix csv name and OwnNamespace InstallModeType
Operator-sdk requires base csv name and operator package name to be the
same for `make bundle` to work.
Update csv name to hw-event-proxy-operator to be consistent with the
package name. When working correctly, `make bundle` synced the
InstallModeType values for the bundle and fixed the issue for OwnNamespace
incorrectly set to false.

Signed-off-by: Jack Ding <jackding@gmail.com>